tests/wpt/web-platform-tests/app-history/navigate/disambigaute-goto-forward-multiple.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));
  assert_equals(appHistory.entries().length, 1);
  assert_equals(i.contentWindow.appHistory.entries().length, 1);
  let initial_key = appHistory.current.key;
  await appHistory.navigate("#top1").committed;
  await appHistory.navigate("#top2").committed;
  assert_equals(appHistory.entries().length, 3);
  assert_equals(i.contentWindow.appHistory.entries().length, 1);
  await i.contentWindow.appHistory.navigate("#1").committed;
  assert_equals(appHistory.entries().length, 3);
  assert_equals(i.contentWindow.appHistory.entries().length, 2);
  assert_equals(appHistory.current.index, 2);
  assert_equals(i.contentWindow.appHistory.current.index, 1);
  assert_true(appHistory.canGoBack);
  assert_true(i.contentWindow.appHistory.canGoBack);
  let final_key = appHistory.current.key;

  await i.contentWindow.appHistory.back().committed;
  assert_equals(appHistory.current.index, 2);
  assert_equals(i.contentWindow.appHistory.current.index, 0);
  await appHistory.goTo(initial_key).committed;
  assert_equals(appHistory.current.index, 0);
  assert_equals(i.contentWindow.appHistory.current.index, 0);
  // There are 2 joint session history entries containing the top window's
  // final key. Navigate to the nearest one (which navigates only the
  // top window).
  await appHistory.goTo(final_key).committed;
  assert_equals(appHistory.current.index, 2);
  assert_equals(i.contentWindow.appHistory.current.index, 0);
}, "appHistory.goTo() goes to the nearest entry when going forward");
</script>
